Update Player constructor to include position and state init
The Player class constructor now accepts an additional parameter,
`sf::Vector2f pos`, to specify the player's position. It initializes
the `PlayableObject` base class with both the position and the name.

Additionally, the constructor ensures proper state initialization
by calling the `enter` method of the `m_state` object after setting
it to `StandingState`.